btrfs-progs: move tree_id parse logic to parse-utils
so that we can reuse it in elsewhere, such as the tree-stat subcommand
