+++ common/parse-utils.h
@@ -31,5 +31,6 @@ int parse_compress_type(const char *type);
 int parse_qgroupid(const char *str, u64 *qgroupid);
 u64 parse_qgroupid_or_path(const char *p);
 int fls64(u64 x);
+u64 parse_tree_id(const char* str);
 
 #endif
